Address review feedback: move tests into BuildTests.cs
- Move 3 Build instance independence tests into existing BuildTests.cs
- Remove inline commentary from Act/Assert section comments
- Delete separate BuildInstanceIndependenceTests.cs file
